+++ src/main/java/ml/pluto7073/plutoscoffee/config/CoffeeConfig.java
@@ -1,23 +1,16 @@
 package ml.pluto7073.plutoscoffee.config;
 
-import ml.pluto7073.pdapi.config.BaseConfig;
+import ml.pluto7073.plutonium.annotations.BooleanOption;
+import ml.pluto7073.plutonium.config.ClientConfig;
 import ml.pluto7073.plutoscoffee.PlutosCoffee;
 
-public class CoffeeConfig extends BaseConfig {
+public class CoffeeConfig extends ClientConfig {
+
+    public static final CoffeeConfig INSTANCE = new CoffeeConfig();
+
+    @BooleanOption(defaultVal = true) public boolean shouldShowCoffeeBar;
 
     public CoffeeConfig() {
-        super("plutoscoffee", "client", PlutosCoffee.LOGGER);
-    }
-    public boolean shouldShowCoffeeBar() {
-        return getBoolean("shouldShowCoffeeBar");
-    }
-
-    public void setShouldShowCoffeeBar(boolean state) {
-        setBoolean("shouldShowCoffeeBar", state);
-    }
-
-    @Override
-    public void initConfig() {
-        setBoolean("shouldShowCoffeeBar", true);
+        super("plutoscoffee", PlutosCoffee.LOGGER, true);
     }
 }